feat: 학교 이름 렌더링 추가 & 카드 슬라이더 오류 처리
- 배차 완료인 내용이 없으면 카드 슬라이드에서 발생하던 에러 해결
