+++ src/pages/Reservations/components/CardSlider/index.tsx
@@ -2,13 +2,11 @@ import {useEffect, useRef, useState} from "react";
 import {ScheduledDepartures} from "./styles.ts";
 import ArrowRightIcon from "@/components/designSystem/Icons/Reservations/ArrowRightIcon";
 import ArrowLeftIcon from "@/components/designSystem/Icons/Reservations/ArrowLeftIcon";
-import LocationIcon from "@/components/designSystem/Icons/Reservations/LocationIcon";
-import CalenderIcon from "@/components/designSystem/Icons/Reservations/CalenderIcon"
-import ClockIcon from "@/components/designSystem/Icons/Reservations/ClockIcon"
-import ArrowRight2Icon from "@/components/designSystem/Icons/Reservations/ArrowRight2Icon"
 import {TO_HOME_ALLOCATED, TO_SCHOOL_ALLOCATED, users_reservations_interface} from "@/api/query/users";
 import {formatKstDate} from "@/utils/time/formatKstDate.ts";
 import {formatKstTime} from "@/utils/time/formatKstTime.ts";
+import Card, {HiddenCard} from "@/pages/Reservations/components/Card";
+import NullCard from "@/pages/Reservations/components/NullCard";
 
 const PREV = -1;
 const CURRENT = 0;
@@ -59,109 +57,50 @@ export default function CardSlider({contents}: CardSliderProps) {
             <ScheduledDepartures.Slider>
                 <HiddenCard />
                 {
-                    renderCards({
-                        slideCurrent,
-                        slideState,
-                        countRef,
-                        data: contents.map(content => ({
-                            direction: content.direction,
-                            date: content.direction === 'TO_HOME' ? formatKstDate(content.operationInfo.busStopArrivalTime, 1) : formatKstDate(content.expectedArrivalTime, 1),
-                            location: content.busStopName,
-                            time: `${content.direction === 'TO_HOME' ? formatKstTime(content.expectedDepartureTime) : formatKstTime(content.operationInfo.busStopArrivalTime)} 탑승`
-                        }))
-                    })
+                    contents.length === 0 ?
+                        <NullCard /> :
+                        renderCards({
+                            slideCurrent,
+                            slideState,
+                            countRef,
+                            data: contents.map(content => ({
+                                direction: content.direction,
+                                date: content.direction === 'TO_HOME' ? formatKstDate(content.operationInfo.busStopArrivalTime, 1) : formatKstDate(content.expectedArrivalTime, 1),
+                                location: content.busStopName,
+                                time: `${content.direction === 'TO_HOME' ? formatKstTime(content.expectedDepartureTime) : formatKstTime(content.operationInfo.busStopArrivalTime)} 탑승`
+                            }))
+                        })
                 }
             </ScheduledDepartures.Slider>
-            <ScheduledDepartures.SliderControls>
-                {
-                    contents.length > 1 ?
-                        <ScheduledDepartures.SlideButton onClick={slidePrev}>
-                            <ArrowLeftIcon />
-                        </ScheduledDepartures.SlideButton>:
-                        null
-                }
-                <ScheduledDepartures.SlideCounter isCurrent={true}>
-                    {slideCurrent.current + 1}
-                </ScheduledDepartures.SlideCounter>
-                <ScheduledDepartures.SlideCounter>
-                    /
-                </ScheduledDepartures.SlideCounter>
-                <ScheduledDepartures.SlideCounter>
-                    {contents.length}
-                </ScheduledDepartures.SlideCounter>
-                {
-                    contents.length > 1 ?
-                        <ScheduledDepartures.SlideButton onClick={slideNext}>
-                            <ArrowRightIcon />
-                        </ScheduledDepartures.SlideButton> :
-                        null
-                }
-            </ScheduledDepartures.SliderControls>
+            {
+                contents.length > 0 &&
+                    <ScheduledDepartures.SliderControls>
+                        {
+                            contents.length > 1 ?
+                                <ScheduledDepartures.SlideButton onClick={slidePrev}>
+                                    <ArrowLeftIcon />
+                                </ScheduledDepartures.SlideButton>:
+                                null
+                        }
+                        <ScheduledDepartures.SlideCounter isCurrent={true}>
+                            {slideCurrent.current + 1}
+                        </ScheduledDepartures.SlideCounter>
+                        <ScheduledDepartures.SlideCounter>
+                            /
+                        </ScheduledDepartures.SlideCounter>
+                        <ScheduledDepartures.SlideCounter>
+                            {contents.length}
+                        </ScheduledDepartures.SlideCounter>
+                        {
+                            contents.length > 1 ?
+                                <ScheduledDepartures.SlideButton onClick={slideNext}>
+                                    <ArrowRightIcon />
+                                </ScheduledDepartures.SlideButton> :
+                                null
+                        }
+                    </ScheduledDepartures.SliderControls>
+            }
         </ScheduledDepartures.Wrapper>
-    )
-}
-
-/* 종속적인 컴포넌트(모듈) */
-type CardProps = {
-    count: number,
-    toUniversity: boolean,
-    date: string,
-    location: string,
-    time: string,
-    isHidden?: boolean
-    isAbsolute?: boolean
-}
-function Card({count, toUniversity, date, location, time, isHidden = false, isAbsolute = true}: CardProps) {
-
-    return (
-        <ScheduledDepartures.Card style={isHidden ? {visibility: 'hidden'}: undefined} isAbsolute={isAbsolute} count={count}>
-            <ScheduledDepartures.CardHeader>
-                <ScheduledDepartures.Title>
-                    탑승 예정
-                </ScheduledDepartures.Title>
-                <ScheduledDepartures.Direction>
-                    <ScheduledDepartures.DirectionText>
-                        {
-                            toUniversity ? "집" : "학교"
-                        }
-                    </ScheduledDepartures.DirectionText>
-                    <ArrowRight2Icon />
-                    <ScheduledDepartures.DirectionText>
-                        {
-                            toUniversity ? "학교" : "집"
-                        }
-                    </ScheduledDepartures.DirectionText>
-                </ScheduledDepartures.Direction>
-            </ScheduledDepartures.CardHeader>
-            <ScheduledDepartures.CardPunchHole left={true}/>
-            <ScheduledDepartures.CardPunchHole left={false}/>
-            <ScheduledDepartures.CardBody>
-                <ScheduledDepartures.InfoBox>
-                    <CalenderIcon />
-                    <ScheduledDepartures.InfoText>
-                        {date}
-                    </ScheduledDepartures.InfoText>
-                </ScheduledDepartures.InfoBox>
-                <ScheduledDepartures.InfoBox>
-                    <LocationIcon />
-                    <ScheduledDepartures.InfoText>
-                        {location}
-                    </ScheduledDepartures.InfoText>
-                </ScheduledDepartures.InfoBox>
-                <ScheduledDepartures.InfoBox>
-                    <ClockIcon />
-                    <ScheduledDepartures.InfoText>
-                        {time}
-                    </ScheduledDepartures.InfoText>
-                </ScheduledDepartures.InfoBox>
-            </ScheduledDepartures.CardBody>
-        </ScheduledDepartures.Card>
-    )
-}
-
-function HiddenCard() {
-    return (
-        <Card count={0} toUniversity={true} date={'1'} location={'1'} time={'1'} isHidden={true} isAbsolute={false}/>
     )
 }
 
@@ -203,4 +142,3 @@ function renderCards({slideCurrent, slideState, countRef, data}: renderCardsProp
     })
 
 }
-
